[REF] crm_timesheet: extract time control into crm_timesheet_time_control
