+++ crm_timesheet/i18n/pt.po
@@ -17,31 +17,9 @@ msgstr ""
 "X-Generator: Weblate 3.10\n"
 
 #. module: crm_timesheet
-#: model_terms:ir.ui.view,arch_db:crm_timesheet.crm_case_kanban_view_leads
-#: model_terms:ir.ui.view,arch_db:crm_timesheet.view_crm_lead_kanban
-msgid "<i class=\"fa fa-lg fa-play-circle\" title=\"Start\"/>"
-msgstr ""
-
-#. module: crm_timesheet
-#: model_terms:ir.ui.view,arch_db:crm_timesheet.crm_case_kanban_view_leads
-#: model_terms:ir.ui.view,arch_db:crm_timesheet.view_crm_lead_kanban
-msgid "<i class=\"fa fa-lg fa-stop-circle\" title=\"Stop\"/>"
-msgstr ""
-
-#. module: crm_timesheet
 #: model:ir.model,name:crm_timesheet.model_account_analytic_line
 msgid "Analytic Line"
 msgstr "Linha Analítica"
-
-#. module: crm_timesheet
-#: model:ir.model,name:crm_timesheet.model_hr_timesheet_switch
-msgid "Helper to quickly switch between timesheet lines"
-msgstr "Ajudante para alternar rapidamente entre linhas de folhas de horas"
-
-#. module: crm_timesheet
-#: model:ir.model.fields,help:crm_timesheet.field_crm_lead__show_time_control
-msgid "Indicate which time control button to show, if any."
-msgstr ""
 
 #. module: crm_timesheet
 #: model:ir.model,name:crm_timesheet.model_crm_lead
@@ -55,30 +33,8 @@ msgid "Project"
 msgstr "Projeto"
 
 #. module: crm_timesheet
-#: model:ir.model.fields,field_description:crm_timesheet.field_crm_lead__show_time_control
-msgid "Show Time Control"
-msgstr ""
-
-#. module: crm_timesheet
-#: model_terms:ir.ui.view,arch_db:crm_timesheet.crm_case_kanban_view_leads
-#: model_terms:ir.ui.view,arch_db:crm_timesheet.crm_case_tree_view_leads
-#: model_terms:ir.ui.view,arch_db:crm_timesheet.crm_case_tree_view_oppor
-#: model_terms:ir.ui.view,arch_db:crm_timesheet.crm_lead_view_form
-#: model_terms:ir.ui.view,arch_db:crm_timesheet.view_crm_lead_kanban
-msgid "Start work"
-msgstr "Iniciar Trabalho"
-
-#. module: crm_timesheet
-#: model_terms:ir.ui.view,arch_db:crm_timesheet.crm_case_kanban_view_leads
-#: model_terms:ir.ui.view,arch_db:crm_timesheet.crm_case_tree_view_leads
-#: model_terms:ir.ui.view,arch_db:crm_timesheet.crm_case_tree_view_oppor
-#: model_terms:ir.ui.view,arch_db:crm_timesheet.crm_lead_view_form
-#: model_terms:ir.ui.view,arch_db:crm_timesheet.view_crm_lead_kanban
-msgid "Stop work"
-msgstr "Parar Trabalho"
-
-#. module: crm_timesheet
 #: model:ir.model.fields,field_description:crm_timesheet.field_crm_lead__timesheet_ids
 #: model_terms:ir.ui.view,arch_db:crm_timesheet.crm_lead_view_form
 msgid "Timesheet"
 msgstr "Registo de horas"
+
