--- crm_timesheet/wizards/__init__.py
@@ -1,2 +1,0 @@
-from . import hr_timesheet_switch
-from . import hr_timesheet_stop
