+++ crm_timesheet_time_control/README.rst
@@ -1,0 +1,114 @@
+.. image:: https://odoo-community.org/readme-banner-image
+   :target: https://odoo-community.org/get-involved?utm_source=readme
+   :alt: Odoo Community Association
+
+==========================
+CRM Timesheet Time Control
+==========================
+
+.. 
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! source digest: sha256:97f30495bf93b8dc167449e8a81b3c43a0d52e1b998220b72e71cd9e33e66f43
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/license-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Ftimesheet-lightgray.png?logo=github
+    :target: https://github.com/OCA/timesheet/tree/19.0/crm_timesheet_time_control
+    :alt: OCA/timesheet
+.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
+    :target: https://translation.odoo-community.org/projects/timesheet-19-0/timesheet-19-0-crm_timesheet_time_control
+    :alt: Translate me on Weblate
+.. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
+    :target: https://runboat.odoo-community.org/builds?repo=OCA/timesheet&target_branch=19.0
+    :alt: Try me on Runboat
+
+|badge1| |badge2| |badge3| |badge4| |badge5|
+
+This module adds start/stop work buttons on leads/opportunities to
+control timesheet timers, gluing together CRM Timesheet and HR Timesheet
+Time Control.
+
+When stopping a timer started from a lead/opportunity, a wizard asks for
+the work description.
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Usage
+=====
+
+- In pipeline/lead form, kanban and list views you have *Start work* /
+  *Stop work* buttons to control the running timesheet timer.
+- When stopping a timer started from a lead/opportunity, a wizard asks
+  you to describe the work done.
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/OCA/timesheet/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us to smash it by providing a detailed and welcomed
+`feedback <https://github.com/OCA/timesheet/issues/new?body=module:%20crm_timesheet_time_control%0Aversion:%2019.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+-------
+
+* Tecnativa
+
+Contributors
+------------
+
+- `Tecnativa <https://www.tecnativa.com>`__
+
+  - Antonio Espinosa
+  - Rafael Blasco
+  - Javier Iniesta
+  - David Vidal
+  - Vicent Cubells
+  - Jairo Llopis
+  - Carolina Fernandez
+  - Víctor Martínez
+
+- Foram Shah <foram.shah@initos.com>
+
+- `Sygel <https://www.sygel.es>`__
+
+  - Harald Panten
+  - Valentin Vinagre
+  - Roger Sans
+
+- `Studio73, S.L. <https://www.studio73.es>`__
+
+  - Pablo Cortés
+
+Maintainers
+-----------
+
+This module is maintained by the OCA.
+
+.. image:: https://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: https://odoo-community.org
+
+OCA, or the Odoo Community Association, is a nonprofit organization whose
+mission is to support the collaborative development of Odoo features and
+promote its widespread use.
+
+This module is part of the `OCA/timesheet <https://github.com/OCA/timesheet/tree/19.0/crm_timesheet_time_control>`_ project on GitHub.
+
+You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
