+++ crm_timesheet_time_control/__init__.py
@@ -1,0 +1,2 @@
+from . import models
+from . import wizards
