+++ crm_timesheet_time_control/__manifest__.py
@@ -4,20 +4,23 @@
 # Copyright 2019 Tecnativa - Jairo Llopis
 # Copyright 2020 Tecnativa - Pedro M. Baeza
 # Copyright 2023 Tecnativa - Carolina Fernandez
+# Copyright 2026 Studio73 - Pablo Cortés
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 {
-    "name": "CRM Timesheet",
+    "name": "CRM Timesheet Time Control",
+    "summary": "Start/stop timesheet timers from leads/opportunities",
     "category": "Customer Relationship Management",
-    "version": "19.0.2.0.0",
-    "depends": ["crm", "hr_timesheet"],
+    "version": "19.0.1.0.0",
+    "depends": ["crm_timesheet", "hr_timesheet_time_control"],
     "data": [
+        "security/ir.model.access.csv",
+        "views/hr_timesheet_stop_view.xml",
         "views/crm_lead_view.xml",
-        "views/crm_team_view.xml",
-        "views/hr_timesheet_view.xml",
     ],
     "author": "Tecnativa, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/timesheet",
     "license": "AGPL-3",
     "installable": True,
+    "auto_install": True,
 }
