+++ crm_timesheet_time_control/i18n/de.po
@@ -1,0 +1,67 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* crm_timesheet_time_control
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 12.0\n"
+"Report-Msgid-Bugs-To: \n"
+"PO-Revision-Date: 2019-07-11 12:43+0000\n"
+"Last-Translator: Maria Sparenberg <maria.sparenberg@gmx.net>\n"
+"Language-Team: none\n"
+"Language: de\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: nplurals=2; plural=n != 1;\n"
+"X-Generator: Weblate 3.7.1\n"
+
+#. module: crm_timesheet_time_control
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_case_kanban_view_leads
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.view_crm_lead_kanban
+msgid "<i class=\"fa fa-lg fa-play-circle\" title=\"Start\"/>"
+msgstr ""
+
+#. module: crm_timesheet_time_control
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_case_kanban_view_leads
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.view_crm_lead_kanban
+msgid "<i class=\"fa fa-lg fa-stop-circle\" title=\"Stop\"/>"
+msgstr ""
+
+#. module: crm_timesheet_time_control
+#: model:ir.model,name:crm_timesheet_time_control.model_account_analytic_line
+msgid "Analytic Line"
+msgstr "Kostenstellenbuchung"
+
+#. module: crm_timesheet_time_control
+#: model:ir.model,name:crm_timesheet_time_control.model_hr_timesheet_switch
+msgid "Helper to quickly switch between timesheet lines"
+msgstr ""
+
+#. module: crm_timesheet_time_control
+#: model:ir.model.fields,help:crm_timesheet_time_control.field_crm_lead__show_time_control
+msgid "Indicate which time control button to show, if any."
+msgstr ""
+
+#. module: crm_timesheet_time_control
+#: model:ir.model.fields,field_description:crm_timesheet_time_control.field_crm_lead__show_time_control
+msgid "Show Time Control"
+msgstr ""
+
+#. module: crm_timesheet_time_control
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_case_kanban_view_leads
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_case_tree_view_leads
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_case_tree_view_oppor
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_lead_view_form
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.view_crm_lead_kanban
+msgid "Start work"
+msgstr ""
+
+#. module: crm_timesheet_time_control
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_case_kanban_view_leads
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_case_tree_view_leads
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_case_tree_view_oppor
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_lead_view_form
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.view_crm_lead_kanban
+msgid "Stop work"
+msgstr ""
