+++ crm_timesheet_time_control/i18n/pt_BR.po
@@ -1,0 +1,68 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* crm_timesheet_time_control
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 12.0\n"
+"Report-Msgid-Bugs-To: \n"
+"PO-Revision-Date: 2023-10-27 21:55+0000\n"
+"Last-Translator: Adriano Prado <adrianojprado@gmail.com>\n"
+"Language-Team: none\n"
+"Language: pt_BR\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: nplurals=2; plural=n > 1;\n"
+"X-Generator: Weblate 4.17\n"
+
+#. module: crm_timesheet_time_control
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_case_kanban_view_leads
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.view_crm_lead_kanban
+msgid "<i class=\"fa fa-lg fa-play-circle\" title=\"Start\"/>"
+msgstr "<i class=\"fa fa-lg fa-play-circle\" title=\"Iniciar\"/>"
+
+#. module: crm_timesheet_time_control
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_case_kanban_view_leads
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.view_crm_lead_kanban
+msgid "<i class=\"fa fa-lg fa-stop-circle\" title=\"Stop\"/>"
+msgstr "<i class=\"fa fa-lg fa-stop-circle\" title=\"Parar\"/>"
+
+#. module: crm_timesheet_time_control
+#: model:ir.model,name:crm_timesheet_time_control.model_account_analytic_line
+msgid "Analytic Line"
+msgstr "Linha Analítica"
+
+#. module: crm_timesheet_time_control
+#: model:ir.model,name:crm_timesheet_time_control.model_hr_timesheet_switch
+msgid "Helper to quickly switch between timesheet lines"
+msgstr ""
+"Auxiliar para alternar rapidamente entre linhas de apontamentos de horas"
+
+#. module: crm_timesheet_time_control
+#: model:ir.model.fields,help:crm_timesheet_time_control.field_crm_lead__show_time_control
+msgid "Indicate which time control button to show, if any."
+msgstr "Indique qual botão de controle de tempo mostrar, se houver."
+
+#. module: crm_timesheet_time_control
+#: model:ir.model.fields,field_description:crm_timesheet_time_control.field_crm_lead__show_time_control
+msgid "Show Time Control"
+msgstr "Mostrar Controle de tempo"
+
+#. module: crm_timesheet_time_control
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_case_kanban_view_leads
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_case_tree_view_leads
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_case_tree_view_oppor
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_lead_view_form
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.view_crm_lead_kanban
+msgid "Start work"
+msgstr "Iniciar Trabalho"
+
+#. module: crm_timesheet_time_control
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_case_kanban_view_leads
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_case_tree_view_leads
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_case_tree_view_oppor
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.crm_lead_view_form
+#: model_terms:ir.ui.view,arch_db:crm_timesheet_time_control.view_crm_lead_kanban
+msgid "Stop work"
+msgstr "Parar Trabalho"
