+++ crm_timesheet_time_control/models/__init__.py
@@ -1,0 +1,2 @@
+from . import account_analytic_line
+from . import crm_lead
