+++ crm_timesheet_time_control/readme/DESCRIPTION.md
@@ -1,0 +1,6 @@
+This module adds start/stop work buttons on leads/opportunities to
+control timesheet timers, gluing together CRM Timesheet and HR
+Timesheet Time Control.
+
+When stopping a timer started from a lead/opportunity, a wizard asks
+for the work description.
