+++ crm_timesheet_time_control/readme/USAGE.md
@@ -1,0 +1,4 @@
+- In pipeline/lead form, kanban and list views you have *Start work* /
+  *Stop work* buttons to control the running timesheet timer.
+- When stopping a timer started from a lead/opportunity, a wizard asks
+  you to describe the work done.
