+++ crm_timesheet_time_control/views/crm_lead_view.xml
@@ -1,0 +1,154 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <!-- Both -->
+    <record id="crm_lead_view_form" model="ir.ui.view">
+        <field name="name">Add timesheet time controls to lead/opportunity form</field>
+        <field name="model">crm.lead</field>
+        <field name="inherit_id" ref="crm.crm_lead_view_form" />
+        <field name="arch" type="xml">
+            <!-- Main lead form action buttons -->
+            <div name="button_box" position="inside">
+                <button
+                    name="button_start_work"
+                    string="Start work"
+                    type="object"
+                    icon="fa-play-circle text-success"
+                    invisible="show_time_control != 'start'"
+                    class="oe_stat_button"
+                    groups="hr_timesheet.group_hr_timesheet_user"
+                />
+                <button
+                    name="button_end_work"
+                    string="Stop work"
+                    type="object"
+                    icon="fa-stop-circle text-warning"
+                    invisible="show_time_control != 'stop'"
+                    class="oe_stat_button"
+                    groups="hr_timesheet.group_hr_timesheet_user"
+                />
+            </div>
+        </field>
+    </record>
+    <!-- Leads -->
+    <record id="view_crm_lead_kanban" model="ir.ui.view">
+        <field name="name">Add timesheet time controls to lead kanban</field>
+        <field name="model">crm.lead</field>
+        <field name="inherit_id" ref="crm.view_crm_lead_kanban" />
+        <field name="arch" type="xml">
+            <xpath expr="//footer" position="inside">
+                <a
+                    name="button_start_work"
+                    invisible="show_time_control != 'start'"
+                    class="o_kanban_inline_block text-success"
+                    string="Start work"
+                    tabindex="-1"
+                    type="object"
+                    groups="hr_timesheet.group_hr_timesheet_user"
+                >
+                    <i class="fa fa-lg fa-play-circle" title="Start" />
+                </a>
+                <a
+                    name="button_end_work"
+                    invisible="show_time_control != 'stop'"
+                    class="o_kanban_inline_block text-warning"
+                    string="Stop work"
+                    tabindex="-1"
+                    type="object"
+                    groups="hr_timesheet.group_hr_timesheet_user"
+                >
+                    <i class="fa fa-lg fa-stop-circle" title="Stop" />
+                </a>
+            </xpath>
+        </field>
+    </record>
+    <record id="crm_case_tree_view_leads" model="ir.ui.view">
+        <field name="name">Add timesheet time controls to lead tree</field>
+        <field name="model">crm.lead</field>
+        <field name="inherit_id" ref="crm.crm_case_tree_view_leads" />
+        <field name="arch" type="xml">
+            <list position="inside">
+                <button
+                    name="button_start_work"
+                    string="Start work"
+                    tabindex="-1"
+                    type="object"
+                    icon="fa-play-circle text-success"
+                    invisible="show_time_control != 'start'"
+                    class="oe_stat_button"
+                    groups="hr_timesheet.group_hr_timesheet_user"
+                />
+                <button
+                    name="button_end_work"
+                    string="Stop work"
+                    tabindex="-1"
+                    type="object"
+                    icon="fa-stop-circle text-warning"
+                    invisible="show_time_control != 'stop'"
+                    class="oe_stat_button"
+                    groups="hr_timesheet.group_hr_timesheet_user"
+                />
+            </list>
+        </field>
+    </record>
+    <!-- Opportunities -->
+    <record id="crm_case_kanban_view_leads" model="ir.ui.view">
+        <field name="name">Add timesheet time controls to opportunity kanban</field>
+        <field name="model">crm.lead</field>
+        <field name="inherit_id" ref="crm.crm_case_kanban_view_leads" />
+        <field name="arch" type="xml">
+            <xpath expr="//footer" position="inside">
+                <a
+                    name="button_start_work"
+                    invisible="show_time_control != 'start'"
+                    class="o_kanban_inline_block text-success"
+                    string="Start work"
+                    tabindex="-1"
+                    type="object"
+                    groups="hr_timesheet.group_hr_timesheet_user"
+                >
+                    <i class="fa fa-lg fa-play-circle" title="Start" />
+                </a>
+                <a
+                    name="button_end_work"
+                    invisible="show_time_control != 'stop'"
+                    class="o_kanban_inline_block text-warning"
+                    string="Stop work"
+                    tabindex="-1"
+                    type="object"
+                    groups="hr_timesheet.group_hr_timesheet_user"
+                >
+                    <i class="fa fa-lg fa-stop-circle" title="Stop" />
+                </a>
+            </xpath>
+        </field>
+    </record>
+    <record id="crm_case_tree_view_oppor" model="ir.ui.view">
+        <field name="name">Add timesheet time controls to opportunity tree</field>
+        <field name="model">crm.lead</field>
+        <field name="inherit_id" ref="crm.crm_case_tree_view_oppor" />
+        <field name="arch" type="xml">
+            <list position="inside">
+                <button
+                    name="button_start_work"
+                    string="Start work"
+                    tabindex="-1"
+                    type="object"
+                    icon="fa-play-circle text-success"
+                    invisible="show_time_control != 'start'"
+                    class="oe_stat_button"
+                    groups="hr_timesheet.group_hr_timesheet_user"
+                />
+                <button
+                    name="button_end_work"
+                    string="Stop work"
+                    tabindex="-1"
+                    type="object"
+                    icon="fa-stop-circle text-warning"
+                    invisible="show_time_control != 'stop'"
+                    class="oe_stat_button"
+                    groups="hr_timesheet.group_hr_timesheet_user"
+                />
+            </list>
+        </field>
+    </record>
+</odoo>
