+++ crm_timesheet_time_control/views/hr_timesheet_stop_view.xml
