+++ crm_timesheet_time_control/wizards/__init__.py
@@ -1,0 +1,2 @@
+from . import hr_timesheet_stop
+from . import hr_timesheet_switch
